Fixed a compiler warning from stout flags.
Review: https://reviews.apache.org/r/21462

+++ 3rdparty/libprocess/3rdparty/stout/include/stout/flags/flags.hpp
@@ -454,7 +454,7 @@ inline Try<Nothing> FlagsBase::load(
 
   // Update 'argc' and 'argv' if we successfully loaded the flags.
   if (!result.isError()) {
-    CHECK_LE(args.size(), *argc);
+    CHECK_LE(args.size(), (size_t) *argc);
     size_t i = 1; // Start at '1' to skip argv[0].
     foreach (char* arg, args) {
       (*argv)[i++] = arg;
